Moved modal to component
